starting styling
