+++ source/pages/catalog/Catalog.css
@@ -1,0 +1,9 @@
+
+#listing-container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    width: 800px;
+    border-radius: 10px;
+}
